use externalTranspiler before doing diffs for status
Create new function `getRawOutputPostExternal` with same interface
as getRawOutput except that it returns a <Promise<ts.EmitOutput>
instead of a ts.EmitOutput.

This function calls getRawOutut, then passes the result through
`runExternalTranspiler` so that babel, if configured, is applied.

Call/use this new function in `getOutputJsStatus`.

Net result: people using babel are not constantly being told their
JS is Outdated.

Closes #736