Center cursor after open in goToDeclaration
Per #679, when going to a declaration, the cursor would often be at the
bottom of the screen. The cursor should now be moved to the middle of
the screen per a call to TextEditor#scrollToCursorPosition once the open
promise resolves.

Thanks to @dsifford for pointing out the docs for the scroll method.